Changed "unsigned int" to "size_t" for 64 bit compatible pointer arithmetics

+++ DirectShowFilters/MPUrlSourceSplitter/MPUrlSourceSplitter/MPUrlSourceSplitter/OutputPinMetadataBox.cpp
@@ -158,7 +158,7 @@ uint64_t COutputPinMetadataBox::GetBoxSize(void)
   return result;
 }
 
-bool COutputPinMetadataBox::ParseInternal(const unsigned char *buffer, uint32_t length, bool processAdditionalBoxes)
+bool COutputPinMetadataBox::ParseInternal(const unsigned char *buffer, size_t length, bool processAdditionalBoxes)
 {
   if (__super::ParseInternal(buffer, length, false))
   {
@@ -198,9 +198,9 @@ bool COutputPinMetadataBox::ParseInternal(const unsigned char *buffer, uint32_t 
   return this->IsSetFlags(BOX_FLAG_PARSED);
 }
 
-uint32_t COutputPinMetadataBox::GetBoxInternal(uint8_t *buffer, uint32_t length, bool processAdditionalBoxes)
+size_t COutputPinMetadataBox::GetBoxInternal(uint8_t *buffer, size_t length, bool processAdditionalBoxes)
 {
-  uint32_t result = __super::GetBoxInternal(buffer, length, false);
+  size_t result = __super::GetBoxInternal(buffer, length, false);
 
   if (result != 0)
   {
@@ -217,7 +217,7 @@ uint32_t COutputPinMetadataBox::GetBoxInternal(uint8_t *buffer, uint32_t length,
 
     if ((result != 0) && processAdditionalBoxes && (this->GetBoxes()->Count() != 0))
     {
-      uint32_t boxSizes = this->GetAdditionalBoxes(buffer + result, length - result);
+      size_t boxSizes = this->GetAdditionalBoxes(buffer + result, length - result);
       result = (boxSizes != 0) ? (result + boxSizes) : 0;
     }
   }
