+++ DirectShowFilters/MPUrlSourceSplitter/MPUrlSourceSplitter/MPUrlSourceSplitter_Parser_F4M/MPUrlSourceSplitter_Parser_F4M.h
@@ -72,7 +72,7 @@ public:
 
 protected:
   // holds last received length
-  unsigned int lastReceivedLength;
+  size_t lastReceivedLength;
 
   /* methods */
