+++ DirectShowFilters/MPUrlSourceSplitter/MPUrlSourceSplitter/MPUrlSourceSplitter_Protocol_Afhs/MPUrlSourceSplitter_Protocol_Afhs.h
@@ -205,8 +205,8 @@ protected:
   unsigned int lastBootstrapInfoUpdateTime;
 
   // holds last processed size from last store time
-  unsigned int lastProcessedSize;
-  unsigned int currentProcessedSize;
+  size_t lastProcessedSize;
+  size_t currentProcessedSize;
 
   /* methods */
